Optimisation and error shutdown on disconnect
All clients now need to send a "REGISTER:XX" command upon connection to server to allocate a Unique ID per client. This is saved to a map for use.
Now forwards only commands received from user application to the relevant client instead of to all clients based on the "ID" received in the message.
This optimised the system to be more responsive.
Now sends a shutdown command "-1" to all clients if user disconnects.